Provision old json postfilters
