+++ server/pkg/provision/provision.go
@@ -4,12 +4,8 @@ import (
 	"context"
 	"time"
 
-	"github.com/cortezaproject/corteza/server/pkg/errors"
-	"github.com/cortezaproject/corteza/server/pkg/id"
 	"github.com/cortezaproject/corteza/server/pkg/options"
-	"github.com/cortezaproject/corteza/server/pkg/rand"
 	"github.com/cortezaproject/corteza/server/store"
-	"github.com/cortezaproject/corteza/server/system/types"
 	"go.uber.org/zap"
 )
 
@@ -39,6 +35,7 @@ func Run(ctx context.Context, log *zap.Logger, s store.Storer, provisionOpt opti
 
 		// Auto-discoveries and other parts that cannot be imported from static files
 		func() error { return emailSettings(ctx, s) },
+		func() error { return apigwFilters(ctx, log.Named("apigw.filters"), s) },
 		func() error { return authAddExternals(ctx, log.Named("auth.externals"), s) },
 		func() error { return oidcAutoDiscovery(ctx, log.Named("auth.oidc-auto-discovery"), s, authOpt) },
 		func() error { return defaultAuthClient(ctx, log.Named("auth.clients"), s, authOpt) },
@@ -49,59 +46,6 @@ func Run(ctx context.Context, log *zap.Logger, s store.Storer, provisionOpt opti
 			return err
 		}
 	}
-
-	return nil
-}
-
-// defaultAuthClient checks if default client exists (handle = AUTH_DEFAULT_CLIENT) and adds it
-func defaultAuthClient(ctx context.Context, log *zap.Logger, s store.AuthClients, authOpt options.AuthOpt) error {
-	if authOpt.DefaultClient == "" {
-		// Default client not set
-		return nil
-	}
-
-	c := &types.AuthClient{
-		ID:     id.Next(),
-		Handle: authOpt.DefaultClient,
-		Meta: &types.AuthClientMeta{
-			Name: "Corteza Web Applications",
-		},
-		ValidGrant: "authorization_code",
-		RedirectURI: func() string {
-			// Disabling protection by redirection URL for now, it caused too much confusion on simple setups
-			//baseURL, _ := url.Parse(authOpt.BaseURL)
-			//return fmt.Sprintf("%s://%s", baseURL.Scheme, baseURL.Hostname())
-			return ""
-		}(),
-
-		Secret:    string(rand.Bytes(64)),
-		Scope:     "profile api",
-		Enabled:   true,
-		Trusted:   true,
-		Security:  &types.AuthClientSecurity{},
-		Labels:    nil,
-		CreatedAt: *now(),
-	}
-
-	_, err := store.LookupAuthClientByHandle(ctx, s, c.Handle)
-	if err == nil {
-		return nil
-	}
-
-	if !errors.IsNotFound(err) {
-		return err
-	}
-
-	if err = store.CreateAuthClient(ctx, s, c); err != nil {
-		return err
-	}
-
-	log.Info(
-		"Added OAuth2 client",
-		zap.String("name", c.Meta.Name),
-		zap.String("redirectURI", c.RedirectURI),
-		zap.Uint64("clientId", c.ID),
-	)
 
 	return nil
 }
